# Add Etag middleware