@kiyonlin 2020-10-15
Remove `generateEtag` helper function:

Old
```hs
Benchmark_Etag
Benchmark_Etag-8         3003379               399 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3014714               407 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3022575               396 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         2949687               407 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3005836               393 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3076695               398 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         2955390               405 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         2992062               403 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3010214               402 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3033276               407 ns/op               0 B/op          0 allocs/op
```

New
```hs
Benchmark_Etag
Benchmark_Etag-8         3051501               390 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3027162               384 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3120348               393 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3175465               394 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         2996120               404 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3082192               392 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3029109               387 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3200316               389 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         3135733               396 ns/op               0 B/op          0 allocs/op
Benchmark_Etag-8         2902838               380 ns/op               0 B/op          0 allocs/op
```