@Fenny 2020-10-15
Linking https://github.com/gofiber/fiber/issues/909